Move BorrowModal into folder
