+++ src/lib/things/BorrowModal/index.ts
@@ -1,0 +1,2 @@
+export { default as BorrowModal } from './BorrowModal.svelte';
+export * from './stores';
